+++ src/provider/aws/ec2/xlb/xlb.ts
@@ -1,0 +1,13 @@
+import { Argv } from 'yargs'
+
+export const command = 'xlb'
+export const desc =
+  'classic, application, and network load-balancer scans (xlb)'
+
+const opts = {
+  exclude: /.*.test.js/,
+}
+
+export const builder = (yargs: Argv) => {
+  return yargs.commandDir('cmds', opts).demandCommand(1)
+}