+++ CHANGELOG.md
@@ -7,11 +7,37 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- AWS - EC2 - xLB - DesyncMitigationMode
+- AWS - EC2 - xLB - WafEnabled
+- AWS - EC2 - xLB - AccessLogsEnabled
+
+## [1.8.0] - 2020-11-11
+
+### Added
+
+- AWS - DynamoDB - EncryptedAtRest
+- AWS - EC2 - ELB - DesyncMitigationMode
+- AWS - EC2 - ELB - AccessLogsEnabled
+- AWS - EC2 - ALB - AccessLogsEnabled
+
+### Changed
+
+### Removed
+
+## [1.7.0] - 2020-11-2
+
+### Added
+
+- Neptune encrypted cluster audit
+- Neptune encrypted instance audit
+
+### Changed
+
 - refactor resource listing
 - `package.json` `main` attribute outputs scan classes
-- adding neptune encrypted cluster audit
-- adding neptune encrypted instance audit
 - fixed bug with async/await in ebs volumeEncrypted
+
+### Removed
 
 ## [1.6.1] - 2020-10-27